testing/web-platform/tests/media-source/URL-createObjectURL-revoke.html

<!doctype html>
<html>
<head>
  <meta charset='utf-8'>
  <title>Revoking a created URL with URL.revokeObjectURL(url)</title>
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
</head>
<body>
<div id="log"></div>
<script>
async_test(function(t) {
    var mediaSource = new MediaSource();
    var url = window.URL.createObjectURL(mediaSource);
    window.URL.revokeObjectURL(url);
    mediaSource.addEventListener('sourceopen',
                                 t.unreached_func("url should not reference MediaSource."));
    var video = document.createElement('video');
    video.src = url;
    video.addEventListener('error', t.step_func_done(function(e) {
        assert_equals(e.target.error.code,
                      MediaError.MEDIA_ERR_SRC_NOT_SUPPORTED,
                      'Expected error code');
        assert_equals(mediaSource.readyState, 'closed');
    }));
}, "Check revoking behavior of URL.revokeObjectURL(url).");
</script>
</body>
</html>
